Calibration: improved handling of rectification parameters
In wizard's stereo results page, allow user to adjust rectification
settings and see their effect on the displayed test image pair.

Export rectification settings (zero-disparity flag and alpha value),
and import them (if available).

+++ stereo-pipeline/rectification_p.h
@@ -60,7 +60,7 @@ protected:
     cv::Mat map11, map12, map21, map22;
 
     // Rectification options
-    float alpha;
+    double alpha;
     bool zeroDisparity;
 };
